requirements.txt
requests==2.31.0
yfinance==1.3.0
groq>=1.2.0
python-dotenv==1.0.1
pandas==2.2.2
